'object' -> 'class' throughout.
Use the [[singleton]] annotation instead of relying on there being no
identifying attributes (in case we haven't inserted them yet -- not
important in this architecture, amy be so later).

Referential attributes have no <name> or <type>, but are marked with
refers="class-name". They imply the use of the GNAT "with type". New
templates attribute-name, attribute-type to cater for this.

Added '_T' to subclass enumeration literal names.